Created node server, updated sliders
Created a node server and created a new ejs file containing the sliders, making it easier to integrate with nodejs. Also made some changes to sliders to make it work better with different screen sizes. The original sliders.html was left unchanged. The new sliders page code can be found in views and can be seen when running the server and going to localhost:8080/. postgresql must be installed on the machine to run the server.